Narrow type conversion

+++ source/l4_proxy.c
@@ -259,8 +259,8 @@ static int s_increment_read_window_l4_proxy(
         }
     }
 
-    uint64_t new_size = aws_add_size_saturating(size, slot->window_size);
-    uint64_t increment = new_size - slot->window_size;
+    size_t new_size = aws_add_size_saturating(size, slot->window_size);
+    size_t increment = new_size - slot->window_size;
 
     if (increment > 0) {
         aws_channel_slot_increment_read_window(slot, increment);
